refactor(windows): isolate core lifecycle and config generation
